refactor(ui): start feed/ split — move shared types to feed/types.ts
First step of the feed.ts god-file split (tracked under codemem-ug38).
Extracts the four public types — FeedItemMetadata, FeedItem, ItemViewMode,
FeedSummary — into packages/ui/src/tabs/feed/types.ts. feed.ts re-exports
FeedItem and FeedItemMetadata for existing callsite stability and imports
the rest internally; no consumer outside feed.ts has to change.

Mechanical move, zero behavior change. Gate (tsc / lint / 1417 tests)
stays green. feed.ts drops from 1729 to 1483 lines.

Left for follow-up sub-commits (same refactor bead):
 - helpers.ts (mergeMetadata, itemKey, authorLabel, canonicalKind, etc.)
 - pagination.ts (module-local offsets + resetPagination + loadMore)
 - card.tsx (FeedItemCard + FeedItemMenu + FeedViewToggle)
 - skeleton.tsx (FeedSkeletonItem)
 - list.tsx (FeedList)
 - inspector.tsx (ContextInspectorPanel + TraceCandidateGroup)
 - view.tsx (FeedTabView)
 - Eventually: feed.ts becomes a thin index.ts re-export

Each follow-up will verbatim-copy the original implementation rather than
rewriting — the point is navigation, not behavior drift.
